chore(code): decouple onhost modules
