+++ super-agent/src/sub_agent/on_host/mod.rs
@@ -1,5 +1,4 @@
 pub mod builder;
 pub mod command;
-pub mod supervisor;
-
 pub mod sub_agent;
+pub mod supervisor;
