--- super-agent/src/sub_agent/on_host/supervisor/README.md
@@ -1,1 +1,0 @@
-A Supervisor is the component that supervises an on-host command.
