+++ super-agent/src/sub_agent/on_host/supervisor/command_supervisor_config.rs
@@ -1,6 +1,6 @@
 use crate::agent_type::health_config::HealthConfig;
 use crate::context::Context;
-use crate::sub_agent::restart_policy::RestartPolicy;
+use crate::sub_agent::on_host::supervisor::restart_policy::RestartPolicy;
 use crate::super_agent::config::AgentID;
 use std::collections::HashMap;
