+++ super-agent/src/sub_agent/on_host/supervisor/mod.rs
@@ -1,3 +1,4 @@
 pub mod command_supervisor;
 pub mod command_supervisor_config;
 pub mod error;
+pub mod restart_policy;
