+++ super-agent/test/on_host/supervisor.rs
@@ -8,7 +8,7 @@ use newrelic_super_agent::sub_agent::on_host::supervisor::command_supervisor::{
     NotStarted, SupervisorOnHost,
 };
 use newrelic_super_agent::sub_agent::on_host::supervisor::command_supervisor_config::SupervisorConfigOnHost;
-use newrelic_super_agent::sub_agent::restart_policy::RestartPolicy;
+use newrelic_super_agent::sub_agent::on_host::supervisor::restart_policy::RestartPolicy;
 use std::sync::Once;
 
 static INIT_LOGGER: Once = Once::new();
